Added stimulation functionality to nsim.py
The simulation class now includes two main ways to incorporate stimuli.
Added the Stim_List function, which adds a list with the given parameters to the simulation
Added the Stim_Set function, which adds a list of stimuli to the simulation.

These two functions are not orthogonal, but allow the user to make a list of stimuli
with similar features to the simulation. They can also manually create their own stimuli
and add that entire list to the simulation. This is for future functionality, where I want
the simulation class to be able to generate lists of neurons with interesting configuration.
This is not necessary however, and may not be incorporated.

Also, the Stimulus class was added. This is a basic encapsulation of the necessary variables
for any given stimulus